requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-crontab==0.7.1
django-js-asset==1.2.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
openpyxl==3.0.7
Pillow==8.4.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
xlwt==1.3.0
